reorg(expansions) split out expandGeom2 expandGeom3 expandPath2
