+++ src/operations/expansions/expand.js
@@ -1,100 +1,10 @@
 const flatten = require('../../utils/flatten')
 
-const { vec2 } = require('../../math')
-
 const { geom2, geom3, path2 } = require('../../geometry')
 
-const { offsetFromGeom2, offsetFromPoints } = require('./offset')
-const area = require('./area')
-const expandShell = require('./expandShell')
-const union = require('../booleans/union')
-
-/*
- * Expand the given geometry (path2) into 2D geometry (geom2).
- * @param {Number} [options.delta=1] - delta (+) of expansion
- * @param {Integer} [options.segments=0] - number of segments when creating rounded corners, or zero for chamfer
- * @returns {geom2} new geometry
- */
-const expandPath2 = (options, geometry) => {
-  const defaults = {
-    delta: 1,
-    segments: 0
-  }
-  let { delta, segments } = Object.assign({ }, defaults, options)
-
-  if (delta <= 0) throw new Error('paths are infinity thin, and cannot contract')
-
-  let points = path2.toPoints(geometry)
-  if (points.length === 0) throw new Error('the given geometry cannot be empty')
-
-  let offsetopts = { delta: delta, segments: segments, closed: geometry.isClosed }
-  let external = offsetFromPoints(offsetopts, points)
-
-  offsetopts = { delta: -delta, segments: segments, closed: geometry.isClosed }
-  let internal = offsetFromPoints(offsetopts, points)
-
-  let newgeometry = null
-  if (geometry.isClosed) {
-    // NOTE: creating path2 from the points insures proper closure
-    let epath = path2.fromPoints({ closed: true }, external)
-    let ipath = path2.fromPoints({ closed: true }, internal.reverse())
-    let esides = geom2.toSides(geom2.fromPoints(path2.toPoints(epath)))
-    let isides = geom2.toSides(geom2.fromPoints(path2.toPoints(ipath)))
-    newgeometry = geom2.create(esides.concat(isides))
-  } else {
-    let capsegments = Math.floor(segments / 2) // rotation is 180 degrees
-    let e2iCap = []
-    let i2eCap = []
-    if (capsegments > 0) {
-      let orientation = area(points)
-      let rotation = orientation < 0 ? -Math.PI : Math.PI
-      let step = rotation / capsegments
-      let eCorner = points[points.length - 1]
-      let e2iStart = vec2.angle(vec2.subtract(external[external.length - 1], eCorner))
-      let iCorner = points[0]
-      let i2eStart = vec2.angle(vec2.subtract(internal[0], iCorner))
-      for (let i = 1; i < capsegments; i++) {
-        let radians = e2iStart + (step * i)
-        let point = vec2.add(eCorner, vec2.scale(delta, vec2.fromAngleRadians(radians)))
-        e2iCap.push(point)
-
-        radians = i2eStart + (step * i)
-        point = vec2.add(iCorner, vec2.scale(delta, vec2.fromAngleRadians(radians)))
-        i2eCap.push(point)
-      }
-    }
-    let allpoints = external.concat(e2iCap, internal.reverse(), i2eCap)
-    newgeometry = geom2.fromPoints(allpoints)
-  }
-  return newgeometry
-}
-
-/*
- * Expand the given geometry.
- * @param {Number} [options.delta=1] - delta (+/-) of expansion
- * @param {Integer} [options.segments=0] - number of segments when creating rounded corners, or zero for chamfer
- * @returns {geom2} new geometry
- */
-const expandGeom2 = (options, geometry) => {
-  const sides = geom2.toSides(geometry)
-  if (sides.length === 0) throw new Error('the given geometry cannot be empty')
-
-  return offsetFromGeom2(options, geometry)
-}
-
-/*
- * Expand the given geometry.
- * @param {Number} [options.delta=1] - delta (+/-) of expansion
- * @param {Integer} [options.segments=defaultResolution3D] - number of segments when creating rounded corners
- * @returns {geom2} new geometry
- */
-const expandGeom3 = (options, geometry) => {
-  const polygons = geom3.toPolygons(geometry)
-  if (polygons.length === 0) throw new Error('the given geometry cannot be empty')
-
-  let expanded = expandShell(options, geometry)
-  return union(geometry, expanded)
-}
+const expandGeom2 = require('./expandGeom2')
+const expandGeom3 = require('./expandGeom3')
+const expandPath2 = require('./expandPath2')
 
 /**
  * Expand the given object(s) using the given options (if any)
